add DEFAULT to -rdoq cli on/off option

+++ Source/API/EbSvtAv1Enc.h
@@ -240,7 +240,7 @@ typedef struct EbSvtAv1EncConfiguration
     /* RDOQ
     *
     * Default is -1. */
-    int8_t                   enable_rdoq;
+    int                      enable_rdoq;
 
     /* Filter intra prediction
     *
